+++ Cheaters/docs/changelog.md
@@ -1,4 +1,8 @@
 # Changelog for Cheaters  #
 
+##v1.1 - 14 April 2013
+* Alleyoop compatible. 
+* Git repo now cloned to `${HOME}/Library/Application Support/Alfred 2/Workflow Data/com.stouty.cheaters/cheaters`
+
 ##v1.0 - 13 April 2013
 * Initial release